requirements.txt
sentence-transformers>=3.0.1
transformers>=4.40.0

python-multipart
aiofiles

docx2txt==0.8
python-docx>=1.1.0

# --- LangChain + LangGraph ---
langchain==1.0.3
langchain-core==1.0.3
langchain-community==0.4.1
langchain-text-splitters==1.0.0
langchain-huggingface==1.0.1
langchain-classic==1.0.0
langgraph>=1.0.2,<1.1.0

faiss-cpu>=1.7.4

# --- API stack ---
fastapi>=0.115,<0.116
uvicorn[standard]>=0.30,<0.31
httpx>=0.27,<0.28

pydantic-settings>=2.3,<3
sqlmodel>=0.0.22,<0.0.23

# elasticsearch==8.15.1
# elastic-transport==8.13.1
